backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-multipart==0.0.20
google-generativeai>=0.8.5
opencv-python==4.11.0.86
pydantic==2.11.5
starlette==0.46.2
typing-extensions==4.13.2
python-dotenv==1.0.0